Added constructors with params for primitives

+++ tcap/tcap-api/src/main/java/org/restcomm/protocols/ss7/tcap/asn/AbortSourceType.java
@@ -32,20 +32,20 @@ public enum AbortSourceType {
 
     User(0), Provider(1);
 
-    private long type = -1;
+    private int type = -1;
 
-    AbortSourceType(long t) {
+    AbortSourceType(int t) {
         this.type = t;
     }
 
     /**
      * @return the type
      */
-    public long getType() {
+    public int getType() {
         return type;
     }
 
-    public static AbortSourceType getFromInt(long t) throws ParseException {
+    public static AbortSourceType getFromInt(int t) throws ParseException {
         if (t == 0) {
             return User;
         } else if (t == 1) {
